Recommend using Windows Terminal instead of mintty
The [Microsoft Terminal](https://github.com/microsoft/terminal) app is a vast improvement over the default Bash for Windows app that ships with WSL distros, and fixes all of the issues with text rendering. Also, the previously recommended mintty for WSL is a f**king mess (it's barely functional and the uninstall batch script is broken. I personally had to spend over an hour reading through it to manually remove all of the files and revert the registry changes it made. I'm still pissed off about it, in case you can't tell).