Remove $ from install cmds to allow for copy paste
This is a pretty minor change. Feel free to close it if you don't agree with it but removing the `$` symbols would expedite installations for users (since you can just copy and paste and the shell runs each command in sequence). As it is now I always have to copy first to an editor, remove the first two columns, and then do a copy paste into a shell.

Cheers.